Implement search and category selection features
Add capability to perform a product search and category selection, which involved changes to the Headerbar, Categories, and productsSlice. The Headerbar now includes a search box that dispatches search actions to the Redux store on input changes. The Categories component dispatches actions to update the current category/subcategory and to fetch new products whenever a new category or subcategory is selected. Lastly, updates to productsSlice include new actions for setting the search field values and category/subcategory, as well as a modified fetchProducts action to include these parameters in the API request.